[GH-130] refactored world menu class
