+++ src/scenes/world-scene.js
@@ -9,7 +9,8 @@ import { getTargetPositionFromGameObjectPositionAndDirection } from '../utils/gr
 import { CANNOT_READ_SIGN_TEXT, SAMPLE_TEXT } from '../utils/text-utils.js';
 import { DialogUi } from '../world/dialog-ui.js';
 import { NPC } from '../world/characters/npc.js';
-import { Menu } from '../world/menu/menu.js';
+//import { Menu } from '../world/menu/menu.js';
+import { WorldMenu } from '../world/world-menu.js';
 import { BaseScene } from './base-scene.js';
 import { DataUtils } from '../utils/data-utils.js';
 import { playBackgroundMusic, playSoundFx } from '../utils/audio-utils.js';
@@ -83,7 +84,7 @@ export class WorldScene extends BaseScene {
   #npcs;
   /** @type {NPC | undefined} */
   #npcPlayerIsInteractingWith;
-  /** @type {Menu} */
+  /** @type {WorldMenu} */
   #menu;
   /** @type {WorldSceneData} */
   #sceneData;
@@ -258,7 +259,7 @@ export class WorldScene extends BaseScene {
     this.#dialogUi = new DialogUi(this, 1280);
 
     // create menu
-    this.#menu = new Menu(this);
+    this.#menu = new WorldMenu(this);
 
     this.cameras.main.fadeIn(1000, 0, 0, 0, (camera, progress) => {
       if (progress === 1) {
