+++ src/world/world-menu.js
@@ -1,0 +1,21 @@
+import { Menu } from '../common/menu/menu.js';
+
+/**
+ * @typedef {keyof typeof MENU_OPTIONS} MenuOptions
+ */
+
+/** @enum {MenuOptions} */
+export const MENU_OPTIONS = Object.freeze({
+  MONSTERDEX: 'MONSTERDEX',
+  MONSTERS: 'MONSTERS',
+  BAG: 'BAG',
+  SAVE: 'SAVE',
+  OPTIONS: 'OPTIONS',
+  EXIT: 'EXIT',
+});
+
+export class WorldMenu extends Menu {
+  constructor(scene) {
+    super(scene, [MENU_OPTIONS.MONSTERS, MENU_OPTIONS.BAG, MENU_OPTIONS.SAVE, MENU_OPTIONS.EXIT]);
+  }
+}
